ai-chat-platform/requirements.txt
# Основные зависимости
flask>=2.3.0
requests>=2.31.0
flask-cors>=4.0.0

# Firebase интеграция
firebase-admin>=6.2.0

# AI провайдеры
openai>=1.3.0
anthropic>=0.8.0
google-generativeai>=0.3.0

# Платежи
stripe>=7.0.0

# Утилиты
python-dotenv>=1.0.0
gunicorn>=21.2.0
